Move Token model into h.models package
As part of an effort to move all models into the h.models package, this
moves the Token model out of h.auth and into h.models.
